requirements.txt
matplotlib==3.5.1
numpy==1.19.2
PyYAML==6.0
scikit_learn==1.1.1
torch==1.11.0
tqdm==4.64.0
